+++ hoka/CHANGELOG.md
@@ -74,6 +74,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Initial  revision.
   
 ---  
-_Copyright 2017-2019 MicroEJ Corp. All rights reserved._  
+_Copyright 2017-2020 MicroEJ Corp. All rights reserved._  
 _This library is provided in source code for use, modification and test, subject to license terms._  
 _Any modification of the source code will break MicroEJ Corp. warranties on the whole library._